Convert to two step form
