+++ components/Modal/styles/modalStyles.ts
@@ -1,6 +1,6 @@
 import React from 'react';
 
-export const modalDiv: React.CSSProperties = {
+export const modalContentContainerStyles: React.CSSProperties = {
   position: 'absolute',
   top: '50%',
   left: '50%',
@@ -10,4 +10,19 @@ export const modalDiv: React.CSSProperties = {
   borderRadius: '4px',
   width: '80%',
   maxWidth: '400px',
+};
+
+export const modalHeaderStyles = {
+  display: 'flex',
+  justifyContent: 'flex-end',
+  marginBottom: '0px',
+};
+
+export const closeModalButtonStyles = {
+  color: 'black',
+  boxShadow: 'none',
+  position: 'absolute',
+  top: '0px',
+  right: '-5px',
+  padding: 0,
 };
