+++ pages/books.tsx
@@ -11,7 +11,7 @@ import React, { useMemo, useState } from 'react';
 import { Box } from '@mui/material';
 
 // custom components
-import AddBook from '@/components/BookCatalog/AddBook';
+import AddBook from '@/components/BookCatalog/AddBookFlow/AddBookModal';
 import CheckInOrOut from '@/components/BookCatalog/CheckInOrOut';
 import Layout from '@/components/Layout';
 import Search from '@/components/Search';
